Fix Receiver panic when querying uninitialized TSDBs
Receivers currently panic when retrieving labels for a TSDB that is
being initialized. The reason for this is that the tenant is added to
the tenants map and the TSDB is started in the background.
When retrieving labels, the MultiTSDB creates Clients for each TSDB,
even if a TSDB is not yet ready.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>