Potential fix for e2e flake from 6114
This commit is a potential fix for the flake e2e test intruduced
in #6114. The commit uses strict endpoints in queriers to make sure
all stores are ready in order for the query to be considered successful.

I verified the fix by running the test locally using
go test ./test/e2e/... -run TestDistributedQueryExecution -count 20.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>