Fix error handling in gRPC query API
The gRPC query api does not propagate errors from the promql engine
to the upstream client. This causes calls against it to fail silently
without any error messages. Noticed while testing the distributed
query mode.

This commit changes the API to propagate errors and warnings as
warn messages. The client can then decide whether to abort on each
message type, or to proceed with execution and present a partial
response message to the user.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>