Remove global lock when pruning tenants
We sometimes see receiver OOMs when pruning tenants. I suspect this is
due to the global lock we acquire during pruning which causes ingestion
for all tenants to stop for a brief amount of time.

This commit replaces the global write lock with a read lock, and
acquires tenant specific locks in places where they are least disruptive.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>