Remove tenant shipper after prunning
We noticed another edge case with tenant prunning where the shipper would
keep syncing blocks even after the tenant TSDB was removed. The reason for this
is because the shipper runs in parallel in a different goroutine and is not
stopped when the TSDB is gone.

This leads to empty shipper files being created on disk once a tenant is pruned,
and the orphaned TSDBs alert from the mixing starts to fire.

This commit modifies the pruning logic to remove all components from a tenant
once eviction conditions have been met.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>