Retry distributed queries
Distributed queries would sometimes fail in a single region leading
to the entire query failing completely. It seems that issues are mostly
related to memory since regional queriers request half as much memory
as global ones.

Before increasing memory requests in NA regions, this commit adds retries
to individual query fragments. This way only parts of the query will be
retried for 3 times before returning an error to the user.