Proxy: Only generate debug messages in debug mode
In debug mode we log information about which endpoints we skip and
query, including the labels of each endpoint. These debug messages, can
use significant memory, especially in larger setups with larger number
of labelsets.

Previously we would generate the debug messages, even when not in debug
mode. With commit, we only generate the debug messages when debug
logging is enabled.

This fixes: https://github.com/thanos-io/thanos/issues/6198

Signed-off-by: Jacob Baungard Hansen <jachanse@redhat.com>