Cache series matchers
Creating Prometheus matchers from protobuf matchers is a CPU intensive
process when the matcher includes a regular expression.

This commit adds a matchers cache from protobuf matcher to Prometheus
matcher which is used to reduce the time spent doing this conversion.

For the time being, the cache has a TTL of 5 minutes, which should be
sufficient to cover most alerting scenarios.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>