Track failed connections to strict endpoints
The EndpointSet will ignore strict endpoints if it fails to connect to them.
This can lead to queries not reporting partial response if one or more such
endpoints are not available when the querier starts.

This commit modifies the EndpointSet to track failed connections separately
from endpoint refs. If one or more connect failures are present at query time,
we inject a store client that will always return an error. This way the query
will either fail, or return a warning when partial response is enabled.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>