Allow customizing the shipper metadata file name
Currently, the shipper metadata file is always called `thanos.shipper.json`
in the Prometheus data directory. This precludes running multiple sidecars
that upload to different object stores, as they will overwrite each other's
metadata file.

This commit allows the metadata file name to be customized via a flag. The
default is unchanged, but it can be overridden with the `--shipper.meta-file-name`
flag.

As part of this, we update the signatures of `WriteMetaFile` and `ReadMetaFile`
to take the full path of the metadata file, rather than just the directory,
and updates the tests that go along with this.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>