compactor: hold lock for a shorter amount of time
If we are constantly running compactor in a loop then we shouldn't pay
the price of constantly holding the lock in the garbage collection
function. What the lock holding means in practice that we have to wait
two or sometimes even three times the amount it takes to sync metas.
That doesn't make sense since we are running the compactor in a loop and
the compacted blocks are properly taken care of.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>