Update Thanos PromQL Engine docs
Move the section on the distributed engine mode into the "Thanos PromQL
Engine" section since the new engine is required for distributed mode.
This also fixes an alignment issue which makes the distributed mode look
like it's part of the Tenancy section.

Also rename the section header to give it clearer "Thanos PromQL Engine"
branding.

Signed-off-by: SuperQ <superq@gmail.com>