Add support for stringlabels in Thanos Query
This commit finalizes support for the stringlabels build tag
so that we can build the binary.

I would assume that we will still get panics if we run a store
since the Series call still relies on casting one pointer type
to another. This will be fixed in a follow up PR.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>