receive: change quorum calculation for RF=2
As discussed during ThanosCon, I am updating the handling for RF=2 to
require only one successful write because requiring all writes to
succeed all the time doesn't make sense and causes lots of confusion to
users. The only other alternative is to forbid RF=2 but I think we
shouldn't do that because people would be forced to add extra resources
when moving from a Sidecar based setup.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>