Optimize zlabel usage for stringlabels
Using the ZLabelsToPromLabels hack is no longer free with stringlabels
because the in-memory format of the two types is different. Because of this,
using a stringlabels build in the Receiver can be more CPU intensive than
the normal build.

In order to allow for an efficient usage of stringlabels, this commit removes
`ZLabelsToPromLabels` calls in TSDBStore and ProxyStore. There is now a specialized
implementation of ZLabel conversion functions with a stringlabels build tag in order to
avoid panics, but long term we should get rid of these functions if possible.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>