thanos-query-frontend: add --query-frontend.query-stats-enabled
We'd like to better instrument which queries are pulling the most data out of Thanos. Tracing helps us dive into the detail, but we'd like to see samples touched at an overall level.

This uses the query-stats feature of thanos-query to append a &stats query parameter to queries, telling thanos-query to return a stats:{} field in the response, then logging the given stats. It is enabled by a --query-frontend.query-stats-enabled flag. Some of this code was originally ported over from cortex-query-frontend but as far as I can tell, never actually tested or used.

Signed-off-by: milesbryant <miles@milesbxf.net>