Fix ExternalLabels() for Prometheus v3.0 (#7893)
Prometheus v3.0.0-rc.0 introduces a new scrape protocol
(`PrometheusText1.0.0`) which is present by default in the global
configuration. It breaks the Thanos sidecar when it wants to retrieve
the external labels.

This change replaces the use of the Prometheus `GlobalConfig` struct by
a minimal struct which unmarshals only the `external_labels` key.

See also https://github.com/prometheus-operator/prometheus-operator/issues/7078

Signed-off-by: Simon Pasquier <spasquie@redhat.com>