pkg/ui/react-app/public/index.html: Fix manifest.json for authenticated UI
We are running Thanos behind a reverse proxy that provides OAuth authentication,
in our specific case using a Vouch proxy which works with cookies.

We have observed the requests to the manifest.json file to fail with a CORS
error. This problem happened also in Prometheus and they fixed it [1], so we are
applying here the same solution.

1: https://github.com/prometheus/prometheus/commit/d30f202c08a7bf4109f18c755c2cbc6a067666bb

Signed-off-by: Vicente Olivert Riera <vincent.olivert.riera@linecorp.com>