receive: Add retry-after backoff with jitter via header field to active-series-limiting
By default, the active-series-limiting does only return a 429 when the
tenant exceeded the limit. Prometheus does parse the retry-after header
by default¹ and if the header is not set, it will do not set a default
backoff on prometheus side. This results in a instant retry of sending
the data again which may result in a potential increase of requests.

¹https://github.com/prometheus/prometheus/blob/bfbae39931a6ddeb5913a6ea8d48a34bbebc6d29/storage/remote/client.go#L309

Signed-off-by: roth-wine <git@roth-wine.de>