feat: tri-state tenant enforcement mode
This turns the current bool config value into an enum that holds the values "false" (former false), "true" (former true), and "soft" which did not exist previously. This allows us to handle soft tenancy use cases that only want to add tenant labels in queries if the tenant header does contain a non-empty value. If the tenant header contains an empty value, no tenant label should be added to a query.

Signed-off-by: Sebastian Hoß <seb@xn--ho-hia.de>