block: bust cache if modified timestamp differs
In the parquet converter, we mark the original meta.json file with a
flag when it gets converted so that Thanos Store wouldn't load it. For
that to work, we need to bust the local cache when that happens.

For tests, we need the updated objstore module so I am doing that as
well.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>