compact: ignore blocks with deletion mark in partial deletes
Blocks deletion always starts with meta.json so if there are multiple
shards of the compactor then one shard can also start to try to delete
the same block because it detects them as partial. Hence, ignore
deletion marks in the partial block cleaning function because blocks
with deletion mark as handled in the other flow.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>