compact: move GC to just after marked block deletion
I noticed in prod that there's a dependency in the garbage collection
function - if we are deleting blocks marked for deletion and a sync
happens in the middle of the garbage collection, it can accidentally
mark the same block for deletion again. Avoid this by passing a slice of
ULIDs that we just deleted.

The effect of not doing this is that we could get errors during querying.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>