block: assume that we do not unmark a block for deletion
Just like we assume that the meta.json file doesn't change, let's also
assume that we do not unmark a block for deletion.

This solves a critical issue in Thanos Store where there is a race
between deletion in compactor and the loading of a block:
- Deletion starts from meta.json, deletion marker are deleted at the end
- Store sees that block, loads it by using the local in-memory and disk
  cache
- By the time the deletion marker filtering functions is executed, the
  marker is deleted by Compactor
- Store happily tries to load that block

The root cause is that we are doing listing & checking markers in two or
more separate steps. Since that is inevitable, we need to assume that
the marker won't disappear until the block is there. This is the
case when everything is working normally.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>