test: fix TestRoundTripQueryCacheWithShardingMiddleware (#8472)
This commit refactors the TestRoundTripQueryCacheWithShardingMiddleware
test to make it more reliable. The previous implementation had a design
flaw because it expected that the PromQL sharding middleware
would always issue 2 HTTP requests per attempt which isn't always the case
(e.g. the first HTTP request might fail before the second request is
issued).

I've run the test on my local machine with `-count 1000` with different
values of GOMAXPROCS (1, 2, 12) and it never failed.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>