cli: Add 'thanos tools' 'query range' and 'query instant' subcommands
Add commands 'thanos tools query range' and 'thanos tools query instant'
that use the Thanos gRPC API to execute PromQL on a Thanos Query
endpoint.

This works like 'promtool query' for the Thanos gRPC API instead of the
Prometheus HTTP API.

Using the Thanos gRPC API improves performance and lowers memory use by
using protobuf instead of text json serialization. It also streams large
result sets to the client, so the whole serialised result does not have
to be accumulated in memory before any of it can be sent; this lowers
latency and peak memory use.

Results cannot be fully streamed, because the Thanos executor still
accumulates them into a promql.Result before sending any. But it helps
reduce the memory overhead and latency of building the whole serialised
response in memory while still holding the result data in memory. This
lowers peak Query memory, and latency to first response.

Sending queries to Thanos Query Frontend is not currently supported,
because Frontend only exposes the Prometheus HTTP API (and only uses the
Prometheus HTTP API to talk to its downstreams).

This is a WIP, with the following improvements still required:

* Needs docs updates
* Needs test cases
* Needs (m)TLS configuration support
* Needs http_proxy support
* Checks for gRPC endpoint format needed (support dns:, ipv4:, etc)
* Needs formatting options like promtool
* --query should be positional parameter

Example use:

   ./thanos tools query range --start='-2h' --step '30s' --insecure \
       --server 'thanos-query.monitoring.svc.local:10901' \
       --timeout=5m --query.promql-engine=thanos \
       --query 'last_over_time({job="kubelet"}[1m])'