Store: *breaking :warning:* set debug.advertise-compatibility-label to false by default (#8509)
In a previous change back in v0.37.0[^1], we stopped explicitely
checking for the compatibility label advertised by the store gateway
introduced to ensure compatibility with querier <v0.8.0.

However, what didn't happen in that PR is that the thanos will still try
to add this label by default[^2].

This change modifies the command line argument to default to NOT adding
this compatibility label by default anymore, since this is now the
assumption made in code.

The reason behind this change is that while upgrading from v0.36.x, we
quickly realised that querries were fanning out to all our
endpoint-groups, and not abiding by our external labels anymore.

[^1]: [pull/7645](https://github.com/thanos-io/thanos/pull/7645/files#diff-fb8e209f24ab370a81b6909a08fd58b18b6dd4a7816597879c08cf5590542b60R800)
[^2]: [debug.advertise-compatibility-label](https://github.com/thanos-io/thanos/blob/d997eed1749e4d91243efc0137f6a20cb659841b/cmd/thanos/store.go#L182-L183)

Signed-off-by: Alexis Lesieur <Xaelias@users.noreply.github.com>