query: use proper labelset for fan-out pruning in distributed engine
To facilitate proper fan-out pruning in the distributed engine, we need
to annotate the remote engines with their proper external labelset.
Previously we only would use their partition labels but that leads to
unnecessary global fan-outs, but newer versions of promql-engine can
distinguish between partition labels and the full labelset.

Signed-off-by: Michael Hoffmann <mhoffmann@cloudflare.com>