receive: consistently mark as unavailable
If we try to write and hear nothing back from the worker (ingester) then
mark that peer as unavailable always. If we hear something back then do
it like previously: map the error code to the status.

Add a test that shows that with artificial timeout the peer gets
removed. The test writes 1000 series. When the randomly generated delay
reaches or exceeds 5s, we can see in the logs that the writing stops for
~5 seconds and then continues succeeding because that lagging peer is
dropped from the hashring.

Loosely based on https://github.com/thanos-io/thanos/pull/8534.

Closing/reconnecting is handled inside of the writer itself.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>